backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.1.0
celery==5.4.0
gunicorn==23.0.0

# ML and NLP
scikit-learn==1.5.2
spacy==3.7.6
nltk==3.9.1
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2

# Web scraping and data processing
beautifulsoup4==4.12.3
scrapy==2.11.2
requests==2.32.3
feedparser==6.0.11

# Development and testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.1

# Production
whitenoise==6.7.0
sentry-sdk==2.14.0